add support for cache tags
