+++ config
@@ -15,7 +15,10 @@ if [ "$HTTP_UWSGI" = "YES" ]; then
 fi
 
 ngx_addon_name=ngx_http_cache_purge_module
-CACHE_PURGE_SRCS="$ngx_addon_dir/ngx_cache_purge_module.c"
+CACHE_PURGE_SRCS="$ngx_addon_dir/ngx_cache_purge_module.c \
+$ngx_addon_dir/ngx_cache_purge_tag.c \
+$ngx_addon_dir/ngx_cache_purge_tag_store_sqlite.c \
+$ngx_addon_dir/ngx_cache_purge_tag_watch_linux.c"
 CORE_LIBS="$CORE_LIBS -lsqlite3"
 
 if [ -n "$ngx_module_link" ]; then
